requirements.txt
pyfredapi
snowflake-connector-python
pandas
python-dotenv
streamlit
trino
pystarburst
plotly
